requirements.txt
Adafruit-DHT==1.3.2
APScheduler==3.3.1
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
bootstrap3-datetime==2.4
certifi==2017.4.17
channels==1.1.6
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
crc16==0.1.1
daphne==1.3.0
Django==1.11.2
django-admin-bootstrapped==2.5.7
django-authtools==1.6.0
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.3
docutils==0.13.1
easy-thumbnails==2.4.1
environ==1.0
gpiozero==1.4.0
gunicorn==19.7.1
html5lib==0.999999999
hyperlink==17.3.0
idna==2.5
image==1.5.5
incremental==17.5.0
msgpack-python==0.4.8
numpy==1.13.0
olefile==0.44
pandas==0.20.2
pbr==3.0.1
Pillow==4.1.1
python-apt==0.9.3.12
python-dateutil==2.6.0
python-redis==0.1.7
pytz==2017.2
redis==2.10.5
requests==2.18.1
RPi.GPIO==0.6.3
six==1.10.0
socketIO-client==0.7.2
South==1.0.2
spidev==3.0
sqlparse==0.2.3
stevedore==1.23.0
Twisted==17.5.0
txaio==2.8.1
tzlocal==1.4
urllib3==1.21.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
webencodings==0.5.1
websocket-client==0.44.0
Werkzeug==0.12.2
zope.interface==4.4.2
